requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.8
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.10.7
packaging==25.0
pluggy==1.6.0
pydantic==2.5.0
pydantic_core==2.18.2
Pygments==2.19.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
typer==0.16.1
typing_extensions==4.14.1
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.30.0
watchfiles==1.1.0
websockets==15.0.1
